# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/certifyInvoices/store": {
            "post": {
                "tags": [
                    "certifyInvoice"
                ],
                "summary": "Create a new Certify Invoice",
                "description": "Create a new Certify Invoice",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "amount",
                                    "date"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "123"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-08-13"
                                    },
                                    "payment_type": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": "2"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": "2"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": "2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ICertifyInvoice"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ICertifyInvoice": {
                "required": [
                    "id",
                    "client_id",
                    "amount",
                    "date",
                    "fac_id",
                    "payment_type",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "amount": {
                        "type": "number",
                        "example": "123"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-13"
                    },
                    "fac_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "payment_type": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-13"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-13"
                    }
                },
                "type": "object"
            },
...
        }
    }
}